Iniciada a rota delete patient. Falta a rota de deletar teste e os testes das rotas feitas.
