+++ public/javascripts/patientDetails.js
@@ -4,21 +4,65 @@ $(document).ready(function() {
   //   location.reload(true);
   // });
 
-  $('#deleteBtn').click(function() {
-    alert("Esse botão vai APAGAR o paciente");
-  });
-
-  // $('#editBtn').click(function() {
-  //   alert("Esse botão vai EDITAR o paciente");
+  // $('#deleteBtn').click(function() {
+  //   alert("Esse botão vai APAGAR o paciente");
   // });
+
+  deletePatient = (patientId) => {
+      $(function() {
+        $( "#dialog-confirm" ).dialog({
+          resizable: false,
+          height: "auto",
+          dialogClass: "no-close",
+          width: 400,
+          modal: true,
+          buttons: {
+            Sim: function() {
+              $( this ).dialog( "close" );
+              $.ajax({
+                method: 'DELETE',
+                url: '/delete/patient/' + patientId,
+                success: successHandler,
+                error: errorHandler
+              });
+            },
+            'Não': function() {
+              $( this ).dialog( "close" );
+            }
+          }
+        });
+      });
+  };
 
   $('#backBtn').click(function() {
     window.history.back();
   });
 
-  // $('#newTestBtn').click(function() {
-  //   alert("Esse botão vai CRIAR NOVO TESTE para o paciente");
-  // });
+  function successHandler () {
+    window.location.replace("/findPatient/");
+
+  }
+
+  function errorHandler (data) {
+    let dialog = $('#dialog-message');
+    dialog.empty();
+    data.responseJSON.error.forEach(error => (dialog.append( "<p>" + error.msg + "</p>" )));
+
+    $( function() {
+      $( "#dialog-message" ).dialog({
+        dialogClass: "no-close",
+        width: 500,
+        title: "Erro ao deletar paciente",
+        modal: true,
+        buttons: {
+          Ok: function() {
+            $( this ).dialog( "close" );
+          }
+        }
+      });
+    } );
+  }
+
 });
