+++ routes/assessment.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const isLoggedIn = require('../config/isLogged');
-const assessmentController = require('../controllers/AssessmentController');
+const AssessmentController = require('../controllers/AssessmentController');
 const validation = require ('../controllers/Validations');
 /*
 Página Inicial
@@ -13,16 +13,16 @@ Página Inicial
   3: Modo para inserção de novo exame para um paciente já existente (trazendo de início os dados do paciente pré-preenchidos) e também permite
      a atualização dos dados do paciente.
 */
-router.get('/', isLoggedIn, assessmentController.assessmentCreateGet);
+router.get('/', isLoggedIn, AssessmentController.assessmentCreateGet);
 // ROTA CREATE - Rota para inserção de um novo paciente com o primeiro exame no banco
-router.post('/', isLoggedIn, validation.validate('assessment'), assessmentController.assessmentCreatePost);
+router.post('/', isLoggedIn, validation.validate('assessment'), AssessmentController.assessmentCreatePost);
 
 //Rota que abre o extrato resumido de um teste para visualização e impressão
-router.get('/assessmentExtract', isLoggedIn, assessmentController.assessmentExtractGet);
+router.get('/assessmentExtract', isLoggedIn, AssessmentController.assessmentExtractGet);
 /*
    Essa rota recebe uma requisição de um teste de um determinado paciente, procura o paciente no banco, e envia os dados pessoais do
 paciente juntamente com os dados do exame requerido
 */
-router.get('/patient/requestAssessment', isLoggedIn, assessmentController.assessmentRequestGet);
+router.get('/patient/requestAssessment', isLoggedIn, AssessmentController.assessmentRequestGet);
 
 module.exports = router;
