+++ views/editPatient/editPatient.ejs
@@ -25,9 +25,6 @@
     <script type="text/javascript" src="/javascripts/controllers/assessmentValidate.js" ></script>
     <script type="text/javascript" src="/javascripts/util.js" ></script>
     <script type="text/javascript" src="/javascripts/editPatient.js" ></script>
-    <script>
-        let patient = jQuery.parseJSON( '<%- patient%>' );
-    </script>
 </head>
 <body>
