+++ views/findPatient/findPatient.ejs
@@ -35,7 +35,7 @@
           <h2 class="text-center border border-dark rounded">Buscar Paciente</h2>
         </div>
       </div>
-      <form id="findPatientsForm" action="findPatient/list" method="POST">
+      <form id="findPatientsForm" action="/findPatient/list" method="POST">
         <div class="row align-items-center"> <%# Nome / data de nascimento %>
           <div class="col-12 col-sm-12 col-md-12 col-lg-8 mx-auto">
             <div class="form-label-group">
