+++ views/findPatient/patientsList.ejs
@@ -62,7 +62,7 @@
             </div>
             <div class="col-12 col-sm-12 col-md-12 col-lg-3 mx-auto">
              <!-- <p class="mb-0">Opções</p>-->
-              <a href="<%=patient._id%>" class="mr-3">
+              <a href="patients/<%=patient._id%>" class="mr-3">
                 <svg xmlns="http://www.w3.org/2000/svg" width="40" height="40" viewBox="0 0 24 24">
                   <path fill="none" d="M0 0h24v24H0V0z"/>
                   <path opacity=".3" d="M12 6c-3.79 0-7.17 2.13-8.82 5.5C4.83 14.87 8.21 17 12 17s7.17-2.13 8.82-5.5C19.17 8.13 15.79 6 12 6zm0 10c-2.48 0-4.5-2.02-4.5-4.5S9.52 7 12 7s4.5 2.02 4.5 4.5S14.48 16 12 16z"/>
